+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added support for passing `Enum` types directly to `make_executable_schema`
 
+
 ## 0.18.1 (2023-02-22)
 
 - Fixed type annotations on `EnumType`.